Refactor new project wizards
Almost all code in both new project and empty project wizards
are identical. So extract common code into an abstract class.
